ObjectAnimator.ofFloat(child, "translationX", /*child.getX(),*/ bean.translationX)
ObjectAnimator.ofFloat(child, "translationX", child.getX(), bean.translationX)

两句等价

+++ transitionhelper/src/main/java/immortalz/me/library/bean/InfoBean.java
@@ -15,8 +15,8 @@ public class InfoBean<ModelType> {
     public Bitmap bitmap;
     private ModelType load;
 
-    public int translationY;
-    public int translationX;
+    public float translationY;
+    public float translationX;
     //origin view
     public Rect originRect = new Rect();
     public int originWidth;
